Separate custom logger tests
